pyproject.toml
[tool.poetry]
name = "django-stripe"
version = "0.1.0"
description = ""
authors = ["Nikita Solouyou <solovev.nikita.05@gmail.com>"]
readme = "README.md"
packages = [{include = "django_stripe"}]

[tool.poetry.dependencies]
python = "^3.11"
django-environ = "^0.11.2"
django = "^4.2.5"
stripe = "^6.5.0"
django-rest-framework = "^0.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
ruff = "^0.0.290"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
